+++ extern/tiff/sources/tif_lfind.c
@@ -1,6 +1,8 @@
 #if defined (ANDROID) || defined (_WIN32_WCE)
 
-void * lfind (const void *key, const void *base, unsigned int *num, unsigned int width, int (*compare)(const void *, const void *))
+# include <sys/types.h>
+
+void * lfind (const void *key, const void *base, size_t *num, size_t width, int (*compare)(const void *, const void *))
 {
   unsigned int place = 0;
